+++ Cargo.toml
@@ -56,9 +56,11 @@ members = [
 pallet-bridge-dispatch = { path = "../darwinia-messages-substrate/modules/dispatch" }
 pallet-bridge-grandpa = { path = "../darwinia-messages-substrate/modules/grandpa" }
 pallet-bridge-messages = { path = "../darwinia-messages-substrate/modules/messages" }
+pallet-bridge-parachains = { path = "../darwinia-messages-substrate/modules/parachains" }
 bp-message-dispatch =  { path = "../darwinia-messages-substrate/primitives/message-dispatch" }
 bp-runtime =  { path = "../darwinia-messages-substrate/primitives/runtime" }
 bp-messages =  { path = "../darwinia-messages-substrate/primitives/messages" }
+bp-polkadot-core = { path = "../darwinia-messages-substrate/primitives/polkadot-core" }
 bridge-runtime-common = { path = "../darwinia-messages-substrate/bin/runtime-common" }
 
 # The list of dependencies below (which can be both direct and indirect dependencies) are crates
