+++ node/runtime/pangolin/src/bridges_message/pangoro.rs
@@ -71,6 +71,10 @@ pub type ToPangoroMessageVerifier =
 /// Encoded Pangolin Call as it comes from Pangoro.
 pub type FromPangoroEncodedCall = FromBridgedChainEncodedMessageCall<Call>;
 
+/// Call-dispatch based message dispatch for Pangoro -> Pangolin messages.
+pub type FromPangoroMessageDispatch =
+	FromBridgedChainMessageDispatch<WithPangoroMessageBridge, Runtime, Ring, WithPangoroDispatch>;
+
 /// The s2s backing pallet index in the pangoro chain runtime.
 pub const PANGORO_S2S_BACKING_PALLET_INDEX: u8 = 20;
 
@@ -302,64 +306,6 @@ impl SourceHeaderChain<<Self as ChainWithMessages>::Balance> for Pangoro {
 		target::verify_messages_proof::<WithPangoroMessageBridge, Runtime, WithPangoroGrandpa>(
 			proof,
 			messages_count,
-		)
-	}
-}
-
-/// Call-dispatch Pangoro -> Pangolin messages.
-#[derive(RuntimeDebug, Clone, Copy)]
-pub struct FromPangoroMessageDispatch;
-impl MessageDispatch<bp_pangolin::AccountId, bp_pangolin::Balance> for FromPangoroMessageDispatch {
-	type DispatchPayload = FromPangoroMessagePayload;
-
-	fn dispatch_weight(
-		message: &DispatchMessage<Self::DispatchPayload, bp_pangolin::Balance>,
-	) -> frame_support::weights::Weight {
-		message.data.payload.as_ref().map(|payload| payload.weight).unwrap_or(0)
-	}
-
-	fn dispatch(
-		relayer_account: &bp_pangolin::AccountId,
-		message: DispatchMessage<Self::DispatchPayload, bp_pangolin::Balance>,
-	) -> MessageDispatchResult {
-		let message_id = (message.key.lane_id, message.key.nonce);
-		pallet_bridge_dispatch::Pallet::<Runtime, WithPangoroDispatch>::dispatch(
-			PANGOLIN_CHAIN_ID,
-			PANGORO_CHAIN_ID,
-			message_id,
-			message.data.payload.map_err(drop),
-			|origin, call| match call {
-				// Filter Ethereum transact call
-				Call::Ethereum(darwinia_ethereum::Call::transact { transaction: tx }) =>
-					match origin.caller() {
-						OriginCaller::Ethereum(RawOrigin::EthereumTransaction(id)) => match tx {
-							// Only support legacy transaction now
-							Transaction::Legacy(t) => {
-								let fee = t.gas_limit.saturating_mul(t.gas_limit);
-								let total_payment = fee.saturating_add(t.value);
-
-								// Ensure the relayer has enough balance
-								let derived_substrate_address = <Runtime as darwinia_evm::Config>::IntoAccountId::derive_substrate_address(*id);
-								if <Runtime as darwinia_evm::Config>::RingAccountBasic::account_balance(relayer_account) >= total_payment {
-										// Ensure the derived ethereum address has enough balance to pay for the transaction
-										let _ = <Runtime as darwinia_evm::Config>::RingAccountBasic::transfer(
-											&relayer_account,
-											&derived_substrate_address,
-											total_payment
-										);
-										return Ok(());
-									}
-								Err(())
-							},
-							// Invalid Ethereum transaction type
-							_ => Err(()),
-						},
-						// Invalid call dispatch origin, should return Err.
-						_ => Err(()),
-					},
-				// Do nothing for other calls.
-				_ => Ok(()),
-			},
 		)
 	}
 }
